move files to footprint-dialog
