+++ playwright-tests/footprint-dialog/footprint-insertion.spec.ts
@@ -1,8 +1,8 @@
 import { test, expect } from "@playwright/test"
-import { viewports } from "./viewports"
+import { viewports } from "../viewports"
 
 for (const [size, viewport] of Object.entries(viewports)) {
-  test.describe(`Footprint Selection tests - ${size} viewport`, () => {
+  test.describe(`Footprint Insertion tests - ${size} viewport`, () => {
     let isMobileOrTablet: boolean
 
     test.beforeEach(async ({ page }) => {
@@ -12,18 +12,26 @@ for (const [size, viewport] of Object.entries(viewports)) {
       isMobileOrTablet = page.viewportSize()?.width! <= 768
     })
 
-    test("footprint selection and preview updates", async ({ page }) => {
+    test("inserts footprint into code", async ({ page }) => {
       if (isMobileOrTablet) {
         await page.click('button:has-text("Show Code")')
       }
       await page.click('button:has-text("Insert")')
       await page.click("text=Footprint")
+      await page.fill(
+        'input[placeholder="Enter chip name (e.g., U1)..."]',
+        "U1",
+      )
       await page.getByRole("combobox").click()
       await page.getByRole("option", { name: "ms012" }).click()
-      await expect(
-        page.locator(".rounded-xl.overflow-hidden svg"),
-      ).toBeVisible()
-      await expect(page).toHaveScreenshot(`footprint-preview-${size}.png`)
+      await page.click('button:has-text("Insert Footprint")')
+      await page.waitForSelector('[role="dialog"]', {
+        state: "hidden",
+        timeout: 5000,
+      })
+      await expect(page.locator(".cm-content")).toContainText("<chip")
+      await expect(page.locator(".cm-content")).toContainText('name="U1"')
+      await expect(page).toHaveScreenshot(`footprint-insertion-${size}.png`)
     })
   })
 }
